This is not really used for anything
It's just provided as a way to guarantee reproducability of the results but
github doens't like old lock files so I'm updating it.

I have not test the code with this version, if you're having problems running
the demos either ignore the lock file or revert to an older version.